chore: update llms log [skip-llms]
