+++ agentLogsStore.json
@@ -1,47 +1,1 @@
-
 {}
-=======
-[
-  {
-    "timestamp": "2025-08-07T10:17:20Z",
-    "command": "npx ts-node --compiler-options '{\"module\":\"CommonJS\"}' scripts/docsync-agent.ts",
-    "output": "Error: supabaseUrl is required.",
-    "synced": false,
-    "syncAttemptedAt": "2025-08-07T10:25:55.847Z",
-    "syncError": "supabase env missing",
-    "architectureDocumented": true,
-    "lifecycleDocumented": true,
-    "designPatternsDocumented": true,
-    "reflectionLogged": true,
-    "codexRegistered": true
-  },
-  {
-    "timestamp": "2025-08-07T12:00:00Z",
-    "command": "npx ts-node scripts/uiSnapshot.ts",
-
-    "snapshotTaken": true
-  },
-  {
-    "timestamp": "2025-08-07T11:05:00Z",
-    "command": "npx ts-node scripts/docsync-agent.ts",
-    "output": "GH_PAT is required.",
-    "synced": false,
-    "syncAttemptedAt": "2025-08-07T11:05:00Z",
-    "syncError": "GH_PAT is required.",
-    "architectureDocumented": true,
-    "lifecycleDocumented": true,
-    "designPatternsDocumented": true,
-=======
-    "snapshotTaken": true,
-    "architectureDocumented": true,
-
-    "codexRegistered": true
-  },
-  {
-    "timestamp": "2025-08-07T10:58:40Z",
-    "command": "npx ts-node --compiler-options '{\"module\":\"CommonJS\"}' scripts/audit-agent-metadata.ts",
-    "output": "Agent metadata is valid and docs are in sync.",
-    "metadataAudited": true,
-    "codexRegistered": true
-  }
-]
