+++ components/AgentTooltip.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import Tooltip from './Tooltip';
-import { agents as agentRegistry } from '../lib/agents/registry';
+import { registry as agentRegistry } from '../lib/agents/registry';
 import { AgentName } from '../lib/types';
 import { formatAgentName } from '../lib/utils';
