+++ lib/agents/pickBot.ts
@@ -1,7 +1,8 @@
-import { agents } from './registry';
+import { loadAgents } from './loadAgents';
 import { Matchup, PickResult } from '../types';
 
 export const pickBot = async (matchup: Matchup): Promise<PickResult> => {
+  const agents = await loadAgents();
   const results = await Promise.all(agents.map((a) => a.run(matchup)));
 
   const teams = [matchup.homeTeam, matchup.awayTeam];
