+++ pages/api/logs.ts
@@ -1,5 +1,5 @@
 import type { NextApiRequest, NextApiResponse } from 'next';
-import { agents } from '../../lib/agents/registry';
+import { registry as agentRegistry } from '../../lib/agents/registry';
 import {
   readAgentLog,
   getAllAgentLogs,
@@ -15,7 +15,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         res.status(404).json({ error: 'Log not found' });
         return;
       }
-      const meta = agents.find((a) => a.name === agentId);
+      const meta = agentRegistry.find((a) => a.name === agentId);
       const weightedScore = data.output?.score * (meta?.weight ?? 1);
       res.status(200).json({
         output: data.output || null,
