+++ pages/api/run-predictions.ts
@@ -5,7 +5,7 @@ import path from 'path';
 import { authOptions } from './auth/[...nextauth]';
 import { loadFlow } from '../../lib/flow/loadFlow';
 import { runFlow, AgentExecution } from '../../lib/flow/runFlow';
-import { agents } from '../../lib/agents/registry';
+import { registry } from '../../lib/agents/registry';
 import type { Matchup, AgentOutputs, PickSummary } from '../../lib/types';
 import { logToSupabase } from '../../lib/logToSupabase';
 
@@ -65,7 +65,7 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
 
       const agentScores: Record<string, number> = {};
       for (const name of flow.agents) {
-        const meta = agents.find((a) => a.name === name);
+        const meta = registry.find((a) => a.name === name);
         const result = outputs[name];
         if (!meta || !result) continue;
         scores[result.team] += result.score * meta.weight;
